+++ src/main/java/com/naveen/productsapi/repository/InventoryRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InventoryRepo extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepo extends JpaRepository<Inventory, Long> {
 
     Optional<Inventory> findByProduct(Product product);
 }
